+++ lib/browser.js
@@ -17,7 +17,7 @@ should
   .use(require('./ext/property'))
   .use(require('./ext/error'))
   .use(require('./ext/match'))
-  .use(require('./ext/jquery'))
+  .use(require('./ext/browser/jquery'))
   .use(require('./ext/deprecated'));
 
  module.exports = should;
