+++ lib/ext/browser/jquery.js
@@ -9,7 +9,7 @@
   * From the jasmine-jquery project under the MIT License.
   */
 
-var util = require('../util');
+var util = require('../../util');
 
 module.exports = function(should, Assertion) {
   var $ = this.jQuery || this.$;
